Task complete

+++ src/main/java/web/UserDao/UserDao.java
@@ -8,12 +8,12 @@ public interface UserDao {
 
     List<UserModel> listUsers();
 
-    UserModel showUserById(int id);
+    UserModel showUserById(Long id);
 
     void add(UserModel userModel);
 
     void update(UserModel userModel);
 
-    void delete(int id);
+    void delete(Long id);
 }
