+++ src/main/java/web/UserServise/UserService.java
@@ -7,12 +7,12 @@ import java.util.List;
 public interface UserService {
     List<UserModel> listUsers();
 
-    UserModel showUserById(int id);
+    UserModel showUserById(Long id);
 
     void add(UserModel userModel);
 
     void update(UserModel userModel);
 
-    void delete(int id);
+    void delete(Long id);
 }
